Fixed all the imports and moved Version.py to the Hardawre folder
