+++ src/Hardware/NanoVNA.py
@@ -23,9 +23,9 @@ import serial
 import numpy as np
 from PyQt6.QtGui import QImage, QPixmap
 
-from Hardware.Serial import drain_serial, Interface
-from Hardware.VNA import VNA
-from Version import Version
+from .Serial import drain_serial, Interface
+from .VNA import VNA
+from .Version import Version
 
 logger = logging.getLogger(__name__)
