+++ src/Hardware/NanoVNA_F.py
@@ -18,8 +18,8 @@
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 import logging
 
-from Hardware.NanoVNA import NanoVNA
-from Hardware.Serial import Interface
+from .NanoVNA import NanoVNA
+from .Serial import Interface
 
 logger = logging.getLogger(__name__)
